float fixed

+++ Unit1-01/index.ts
@@ -21,7 +21,7 @@ import { createPrompt } from 'bun-promptx'
 console.log('Choose the length of the logs you would like to move with a truck.')
 const lengthString = createPrompt('Choose a maple log length from among 0.25m, 0.5m, and 1m: ')
 
-const lengthNumber = parseInt(lengthString.value)
+const lengthNumber = parseFloat(lengthString.value)
 
 const logNumber = 55 / lengthNumber
